+++ src/selectedtests/app/evergreen.py
@@ -1,6 +1,3 @@
-"""Helper functions for selected tests API."""
-from typing import List
-
 from evergreen import EvergreenApi, Project
 from fastapi import HTTPException
 
@@ -19,13 +16,3 @@ def try_retrieve_evergreen_project(project: str, api: EvergreenApi) -> Project:
     if not evergreen_project:
         raise HTTPException(status_code=404, detail="Evergreen project not found")
     return evergreen_project
-
-
-def parse_changed_files(changed_files: str) -> List[str]:
-    """
-    Get the list of strings specified in CSV format under change_files query parameter.
-
-    :param changed_files: The CSV string.
-    :return: The parsed list of strings.
-    """
-    return changed_files.split(",")
