+++ tests/selectedtests/app/controllers/test_project_task_mappings_controller.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock, patch
 from miscutils.testing import relative_patch_maker
 from starlette.testclient import TestClient
 
-from selectedtests.app.helpers import __name__ as helpers_ns
+from selectedtests.app.parsers import __name__ as helpers_ns
 
 NS = "selectedtests.app.controllers.project_task_mappings_controller"
